src/requirements.txt
asgiref==3.8.1
attrs==25.1.0
django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
phonenumbers==8.13.55
pip==25.0.1
psycopg2==2.9.10
pyjwt==2.10.1
python-dotenv==1.0.1
pyyaml==6.0.2
referencing==0.36.2
rpds-py==0.22.3
setuptools==75.8.0
sqlparse==0.5.3
typing-extensions==4.12.2
uritemplate==4.1.1
uv==0.6.0
wheel==0.45.1
